requirements.txt
Django==4.2.0
djangorestframework==3.14.0
qrcode==7.4.2
Pillow>=10.0.0
django-cors-headers==4.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
openai==1.3.0
google-generativeai==0.3.1
